pipeline/pyproject.toml
[tool.poetry]
name = "test-pipeline"
version = "0.1.0"
description = ""
authors = ["Cristian Garcia <cgarcia.e88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kfp = "^1.0.4"
typer = "^0.3.2"
matplotlib = "^3.3.2"
numpy = ">=1.16.0,<1.19.0"
tensorflow-cpu = "^2.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
